+++ CHANGELOG.md
@@ -2,12 +2,16 @@
 
 ## New Features
 
+- PR #126 Create and build cuSpatial Docs 
 - PR #130 Add cubic spline fit and interpolation
 
 ## Improvements
 
+- PR #128 Use RMM's `DeviceBuffer` for Python allocations
+
 ## Bug Fixes
 
+- PR #123 Update references to error utils after libcudf changes
 
 
 # cuSpatial 0.12.0 (04 Feb 2020)