+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# cuSpatial 0.14.0 (Date TBD)
+
+## New Features
+
+## Improvements
+
+## Bug Fixes
+
 # cuSpatial 0.13.0 (TBD)
 
 ## New Features